Update README.md
Toy Problems
Challenge 1. Student grading
Write a function that prompts the user to input student marks. The input should be between 0 and 100. The output should correspond the correct grade, as shown below: 

        A > 79, B - 60 to 79, C -  59 to 49, D - 40 to 49, E - less 40.

In this challenge. I wrote a javascript program that prompts the user the user to enter the student mark. Once the mark is entered, it shows the student grade automatically. It uses java conditions of if... else if in executing where if... excutes ablock of code if condition is true. else if...specifies new condition if the first failed. 

Challenge 2. Speed detector.

Write a program that takes the speed of a car as input e.g 80. If the speed is less than 70, it should print “Ok”. Otherwise, for every 5 km/s above the speed limit (70), it should give the driver one demerit point and print the total number of demerit points.

   > For example, if the speed is 80, it should print: “Points: 2”. If the driver gets more than 12 points, the function should print: “License suspended”.

This is a code that tests the speed of the driver. It calculates based on the required speed of 70 Km/hr. Any increase in speed  by 5Km/hr is treated is treated as a demerit. The code will tell the driver that his/her license is suspended if demerits are 12 and beyond. The use of if... and if else... conditions test if either one condition is true like being within the range and if else... when false.

Challenge 3. Net Salary Calculator
 Write a program whose major task is to calculate an individual’s Net Salary by getting the inputs of basic salary and benefits. Calculate the payee (i.e. Tax), NHIF Deductions, NSSF Deductions, gross salary, and net salary. 

NB: Use KRA, NHIF, and NSSF values provided in the link below.

- https://www.aren.co.ke/payroll/taxrates.htm Links to an external site.

 Links to an external site.-  Links to an external site.Links to an external site.

- www.kra.go.ke/en/individual/calculate-tax/calculating-tax/paye

In this challenge. The code perfoms arithmetics so as to achieve the net pay of the employee. Gross salary is calculated as basic salary plus all allowances. Net pay is calculated as Gross salary minus all deductions. In this challenge, the employee was entitled to benefits and also subject to deductions depending on the earnings. The program applies the formula and gives the employee the net pay he/she receives.